Small ui change
The send coin button at the bottom of the main transaction screen takes up more space, because I use that more, and the scan a code to send button that was in the center of the bar was removed, because I don't use that much (you can still choose the camera option on the send coins screen). Shortened request coins to request, since I don't use it that much and it takes up less space.